fix tests
Since #60076 using a closure as the command code that does not return an
integer is deprecated. This means that our tests can trigger deprecations
on older branches when the high deps job is run. This usually is not an
issue as we silence them with the `SYMFONY_DEPRECATIONS_HELPER` env var.
However, phpt tests are run in a child process where the deprecation error
handler of the PHPUnit bridge doesn't step in. Thus, for them deprecations
are not silenced leading to failures.